+++ app/users/dashboard/workforce/invitations/[id]/page.tsx
@@ -8,8 +8,8 @@ import {
   CardDescription,
   CardHeader,
   CardTitle,
-} from '@/components/ui/card';
-import { Button } from '@/components/ui/button';
+} from '@/components/shadcn/card';
+import { Button } from '@/components/shadcn/button';
 import {
   ArrowLeft,
   Mail,
